Implement StateManager
The StateManager  operates as follow:
- A System exposes a set of variable its objects possess in order for observers
to be able to observe them. To expose a variable, one must declare a static
constexpr VarType<S, T> in the struct System::Var. The variables must have
different indexes in order to differentiate them. S must be the System's cla²ss
and T the type of the underlying variable.
- Observer (or anybody else) begins observing a Subject's variable by asking
the StateManager to create a connection between the Observer and the Subject
for a certain variable identified by a (System, ID) pair where the System is
represented by the hash generated by type_index and ID is a user defined ID
set when exposing the variables (see first bullet). The StateManager ultimately
stores a callback it can fire when the observed value changes. The callback
will modify the value of one of the Observer's member that was set to be the
listener when creating the connection.
- An Object may associate an Observable<T> with a VarType<S, T> in order to
make it available to other objects. This is done through a member method.
It is necessary Observable and VarType refer to the same type T. This is
assured at compile time thanks to templates. When the Observable's value is
modified, a callback is fired that informs the StateManager of the new value.
The StateManager stores this information in a queue.
- At the end of a cycle, when it's sync time, the StateManager distributes all
the changes to the Observers concerned. That is those who were observing
variables that were modified during the cycle.